requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
python-multipart==0.0.9
google-cloud-storage==2.14.0
celery[redis]==5.3.6
python-dotenv==1.0.1
pydantic-settings==2.2.1
requests==2.31.0
google-generativeai>=0.7.0
jinja2==3.1.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

python-docx>=1.1.0

fastapi-mail>=1.4.1
